+++ examples/complete.lua
@@ -1,0 +1,5 @@
+-- this complete example cover : 
+-- - functions watch, destroy, unwatch and is_watched
+-- - complete use of events set and get
+
+-- TODO